requirements.txt
Flask
Flask-Session
flask-sqlalchemy>=0.15
SQLAlchemy<1.4.0
markdown
bcrypt
unidecode
Pillow
